Importable tests; start to implement dryrun/verbose.

+++ lib/snapit/__init__.py
@@ -16,6 +16,14 @@ import os.path
 
 DEFAULT_CURRENT_PATH = './current/'
 TIMESTAMP_STRFTIME_FMT = '%Y-%m-%d_%H:%M:%S'
+INFO_VERBOSE = False
+INFO_DRYRUN = False
+
+def info(message):
+    if INFO_VERBOSE:
+        if INFO_DRYRUN:
+            message = 'Would: ' + message
+        print(message)
 
 
 def ensure_path(path):
@@ -23,7 +31,7 @@ def ensure_path(path):
         os.makedirs(path)
 
 
-def make_soft_image(target_dirpath, relative_source_dirpath):
+def make_soft_image(target_dirpath, relative_source_dirpath, dryrun=False):
     """
     Create a duplicate of a directory with relative-path softlinks for all
     content.
@@ -38,19 +46,30 @@ def make_soft_image(target_dirpath, relative_source_dirpath):
         If it does already exist, it must be empty.
     * relative_source_dirpath (string):
         directory to create an image of, as a path relative to the target.
+    * dryrun (bool):
+        don't actually perform (N.B. debug output is controlled by globals).
 
     """
+    actual_source_dirpath = os.path.join(target_dirpath,
+                                         relative_source_dirpath)
+    info('Make soft image of {!r} in {!r}'.format(
+        actual_source_dirpath, target_dirpath))
+
     # Create target if it does not exist.
-    ensure_path(target_dirpath)
+    info('Ensure {!r} exists.'.format(target_dirpath))
+    if not dryrun:
+        ensure_path(target_dirpath)
 
     # Check target is empty.
+    info('Check {!r} is empty.'.format(target_dirpath))
     if len(os.listdir(target_dirpath)) > 0:
         msg = 'Aborted creating links in "{!s}" as it is not empty.'
         raise ValueError(msg.format(target_dirpath))
 
-    from_actual_dirpath = os.path.join(target_dirpath, relative_source_dirpath)
     # Fill target with pointers to the source data.
-    for name in os.listdir(from_actual_dirpath):
+    for name in os.listdir(actual_source_dirpath):
+        info('Create soft-link for {!r} in {!r}'.format(
+            name, actual_source_dirpath))
         target_path = os.path.join(target_dirpath, name)
         source_relative_path = os.path.join(relative_source_dirpath, name)
         source_actual_path = os.path.join(target_dirpath, source_relative_path)
@@ -59,7 +78,8 @@ def make_soft_image(target_dirpath, relative_source_dirpath):
             # Make path to the source subdir relative to the target subdir.
             source_subdir_relative_path = os.path.join(
                 '..', source_relative_path)
-            make_soft_image(target_path, source_subdir_relative_path)
+            make_soft_image(target_path, source_subdir_relative_path,
+                            dryrun=dryrun)
         else:
             # Create a softlink.
             if os.path.islink(source_actual_path):
@@ -69,12 +89,15 @@ def make_soft_image(target_dirpath, relative_source_dirpath):
                 # So a relative path like "../../snap_1/subdir_a/actual_file"
                 # is correct from within *any* "<snapshot>/subdir_a".
                 source_relative_path = os.readlink(source_actual_path)
-            os.symlink(source_relative_path, target_path)
+            info('Create softlink at {!r} to {!r}.'.format(
+                target_path, source_relative_path))
+            if not dryrun:
+                os.symlink(source_relative_path, target_path)
 
 
 def snapshot(current_path=None,
-             snapshot_name=None):
-    #        verbose=None, dryrun=False):
+             snapshot_name=None,
+             verbose=False, dryrun=False):
     """
     Snapshot the contents of a directory tree using softlinks.
 
@@ -97,8 +120,8 @@ def snapshot(current_path=None,
         print summary of actions. Defaults to ='dryrun'.
 
     """
-#    if verbose is None:
-#        verbose = dryrun
+    if verbose is None:
+        verbose = dryrun
 
     if not os.path.exists(current_path):
         msg = 'Current content path, "{!s}", does not exist.'
@@ -110,14 +133,17 @@ def snapshot(current_path=None,
     if snapshot_name is None:
         timestamp = datetime.datetime.now()
         snapshot_name = timestamp.strftime(TIMESTAMP_STRFTIME_FMT)
-
     snapshot_path = os.path.join(parent_path, snapshot_name)
+
     # Rename the current content to the snapshot path.
-    shutil.move(current_path, snapshot_path)
+    info('Move (rename) target {!r} to new snapshot {!r}'.format(
+        current_path, snapshot_path))
+    if not dryrun:
+        shutil.move(current_path, snapshot_path)
 
     # Create a new 'current' full of links to the original content.
     snapshot_relative_path = os.sep.join(['..', snapshot_name])
-    make_soft_image(current_path, snapshot_relative_path)
+    make_soft_image(current_path, snapshot_relative_path, dryrun=dryrun)
 
 
 def _make_parser():
@@ -145,6 +171,9 @@ def _make_parser():
 if __name__ == '__main__':
     parser = _make_parser()
     args = parser.parse_args()
+
+    INFO_VERBOSE = args.verbose or args.dryrun
+    INFO_DRYRUN = args.dryrun
+
     snapshot(current_path=args.current,
              snapshot_name=args.name)
-#             verbose=args.verbose, dryrun=args.dry_run)
